% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{buildScriptBundle}
\alias{buildScriptBundle}
\alias{buildRmdBundle}
\title{Produce a zip bundle of code and results}
\usage{
buildScriptBundle(
  code = NULL,
  output_zip_path,
  script_name = "script.R",
  include_files = list(),
  render = TRUE,
  render_args = list()
)

buildRmdBundle(
  report_template,
  output_zip_path,
  vars = list(),
  include_files = list(),
  render = TRUE,
  render_args = list()
)
}
\arguments{
\item{code}{A language object.}

\item{output_zip_path}{A filename for the resulting zip bundle.}

\item{script_name}{A name for the R script in the zip bundle.}

\item{include_files}{A named list consisting of additional files that should
be included in the zip bundle. The element names indicate the destination
path within the bundle, specified as a relative path; the element values
indicate the path to the actual file currently on disk, specified as either
a relative or absolute path.}

\item{render}{Whether or not to call \code{\link[rmarkdown:render]{rmarkdown::render()}} on the R script.}

\item{render_args}{Arguments to provide to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{report_template}{Filename of an Rmd template to be expanded by \code{\link[knitr:knit_expand]{knitr::knit_expand()}}.}

\item{vars}{A named list of variables passed along to \code{...} in \code{\link[knitr:knit_expand]{knitr::knit_expand()}}.}
}
\description{
Produce a zip bundle of code and results
}
\seealso{
knitr::knit_expand
}
