% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatCode}
\alias{formatCode}
\title{Format code objects}
\usage{
formatCode(code, localize = "auto", unpack = TRUE)
}
\arguments{
\item{code}{Language/expr object (recommended), or character vector}

\item{localize}{whether or not to \code{\link[=local]{local()}}ize assignment calls that
begin with an open brace (that is, \code{\{}). The default, \code{"auto"}, wraps
these calls if they contain a \code{\link[=return]{return()}} statement.}

\item{unpack}{If an assigment call begins with an open brace (that is, \code{\{}),
and the \code{\{} call has not been \code{localize}d, should we move the assignment
to the child of the \code{\{} call?}
}
\value{
Single-element character vector with formatted code
}
\description{
Converts language/expression objects (i.e. quoted code) into pretty-formatted
text. This logic is based on \code{\link[styler:style_text]{styler::style_text()}}, but with additional
features/opinions; see Details.
}
\details{
\code{formatCode} differs from \code{\link[styler:style_text]{styler::style_text()}} in a few ways:
\itemize{
\item Pipe operators (\code{\%>\%}) are always followed by a line break.
\item Superfluous \code{\{} and \code{\}} are removed in many cases.
\item Since quoted R code cannot contain comments, you can use comment strings.
}

Comment strings are literal strings that appear on their own line, and begin
with one or more \code{#} characters. Such strings will be converted into comments
by \code{formatCode}.
}
