% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.shinyMetaExpr}
\alias{knit_print.shinyMetaExpr}
\title{Knitr S3 methods}
\usage{
knit_print.shinyMetaExpr(x, ...)
}
\arguments{
\item{x}{Object to knit_print}

\item{...}{Additional knit_print arguments}
}
\description{
This S3 method allows \code{\link[=metaExpr]{metaExpr()}}s to print themselves in
knitr/rmarkdown documents.
}
