% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareactive.R
\name{metaExpr}
\alias{metaExpr}
\title{Mark an expression as a meta-expression}
\usage{
metaExpr(expr, env = parent.frame())
}
\arguments{
\item{expr}{an expression.}

\item{env}{an environment.}
}
\description{
Mark an expression as a meta-expression
}
\seealso{
\code{\link[=metaReactive2]{metaReactive2()}}, \code{\link[=metaObserve2]{metaObserve2()}}, \code{\link[=metaRender2]{metaRender2()}}
}
