% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{outputCode}
\alias{outputCode}
\title{Overlay an icon on a shiny output}
\usage{
outputCode(outputObj, heading = tags$span(style = "color: #707070",
  "Show code", icon("code")))
}
\arguments{
\item{outputObj}{A shiny output.}

\item{heading}{A shiny \link{tags} object for placing in the header panel.}
}
\description{
Intended for overlaying an icon over a shiny output, that when clicked,
displays or downloads code for reproducing that output. The icon
works similar to an \code{\link[shiny:actionButton]{shiny::actionButton()}}, where an input value,
tied to the icon, increases by one each time it is pressed. The name
of that input is a function of \code{outputObj}'s \code{outputId}:
\code{input$OUTPUTID_shinymeta_icon}.
}
\examples{

if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    sliderInput("n", label = "Number of samples", min = 10, max = 100, value = 30),
    outputCode(plotOutput("p"))
  )
  server <- function(input, output) {
    output$p <- metaRender(renderPlot, {
      plot(sample(!!input$n))
    })
    observeEvent(input$p_shinymeta_icon, {
      code <- expandObjects(output$p)
      displayCodeModal(code)
    })
  }
  shinyApp(ui, server)
}

}
\seealso{
\link{displayCodeModal}
}
