% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shinyMetaExpr}
\alias{print.shinyMetaExpr}
\title{Print method for shinymeta metaExpression}
\usage{
\method{print}{shinyMetaExpr}(x, ...)
}
\arguments{
\item{x}{an object of class shinyMetaExpr.}

\item{...}{arguments passed along to \code{\link[=formatCode]{formatCode()}}.}
}
\description{
Print method for shinymeta metaExpression
}
\keyword{internal}
