% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareactive.R
\name{metaMode}
\alias{metaMode}
\alias{withMetaMode}
\title{Evaluate an expression with meta mode activated}
\usage{
metaMode(value)

withMetaMode(expr, mode = TRUE)
}
\arguments{
\item{expr}{an expression.}

\item{mode}{whether or not to evaluate expression in meta mode.}
}
\description{
Evaluate an expression with meta mode activated
}
\seealso{
\code{\link[=expandCode]{expandCode()}}
}
