% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{displayCodeModal}
\alias{displayCodeModal}
\title{Display a shinyAce code editor via shiny modal}
\usage{
displayCodeModal(code, title = NULL, clip = "clipboard",
  footer = modalButton("Dismiss"), size = c("m", "s", "l"),
  easyClose = TRUE, fade = TRUE,
  session = shiny::getDefaultReactiveDomain(), ...)
}
\arguments{
\item{code}{Either a language object or a character string.}

\item{title}{An optional title for the dialog.}

\item{clip}{An \code{\link[=icon]{icon()}} \code{name} that a user can press to copy \code{code} to the clipboard.
If you wish to not have an icon, specify \code{clip = NULL}.}

\item{footer}{UI for footer. Use \code{NULL} for no footer.}

\item{size}{One of \code{"s"} for small, \code{"m"} (the default) for medium,
or \code{"l"} for large.}

\item{easyClose}{If \code{TRUE}, the modal dialog can be dismissed by
clicking outside the dialog box, or be pressing the Escape key. If
\code{FALSE} (the default), the modal dialog can't be dismissed in those
ways; instead it must be dismissed by clicking on the dismiss button, or
from a call to \code{\link[=removeModal]{removeModal()}} on the server.}

\item{fade}{If \code{FALSE}, the modal dialog will have no fade-in animation
(it will simply appear rather than fade in to view).}

\item{session}{a shiny session object (the default should almost always be used).}

\item{...}{arguments passed along to \code{shinyAce::aceEditor()}}
}
\description{
Show a \code{shinyAce::aceEditor()} in a \code{shiny::modalDialog()}.
}
\examples{

if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    sliderInput("n", label = "Number of samples", min = 10, max = 100, value = 30),
    actionButton("code", icon("code")),
    plotOutput("p")
  )
  server <- function(input, output) {
    output$p <- metaRender(renderPlot, {
      plot(sample(!!input$n))
    })
    observeEvent(input$code, {
      code <- expandObjects(output$p)
      displayEditor(code)
    })
  }
  shinyApp(ui, server)
}

}
\seealso{
\link{outputCodeButton}
}
