% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareactive.R
\name{metaReactive}
\alias{metaReactive}
\title{Create a meta-reactive expression}
\usage{
metaReactive(expr, env = parent.frame(), quoted = FALSE,
  label = NULL, domain = getDefaultReactiveDomain())
}
\description{
Create a reactive that operates in one of two modes, depending on how it is
invoked.
}
\details{
When invoked normally, it operates just like a normal reactive.

When invoked with meta-mode activated (i.e. withMetaMode is on the call
stack), instead of executing the code and returning the value, it returns the
code expression.
}
