% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{metaRender}
\alias{metaRender}
\alias{metaRender2}
\title{Create a meta-reactive output}
\usage{
metaRender(renderFunc, expr, ..., env = parent.frame(), quoted = FALSE,
  localize = "auto", bindToReturn = FALSE)

metaRender2(renderFunc, expr, ..., env = parent.frame(),
  quoted = FALSE)
}
\arguments{
\item{renderFunc}{A reactive output function (e.g., \link[shiny:renderPlot]{shiny::renderPlot}, \link[shiny:renderText]{shiny::renderText}, \link[shiny:renderUI]{shiny::renderUI}, etc).}

\item{expr}{An expression that generates given output expected by \code{renderFunc}.}

\item{...}{Other arguments passed along to \code{renderFunc}.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression.}

\item{quoted}{Is the expression quoted? By default, this is \code{FALSE}.
This is useful when you want to use an expression that is stored in a
variable; to do so, it must be quoted with \code{quote()}.}

\item{localize}{Whether or not to wrap the returned expression in \code{\link[=local]{local()}}.
The default, \code{"auto"}, only wraps expressions with a top-level \code{\link[=return]{return()}}
statement (i.e., return statements in anonymized functions are ignored).}

\item{bindToReturn}{For non-\code{localize}d expressions, should an assignment
of a meta expression be applied to the \emph{last child} of the top-level \code{\{} call?}
}
\description{
Create a meta-reactive output that, when invoked with meta-mode activated
(i.e. called within \code{\link[=expandCode]{expandCode()}} or \code{\link[=withMetaMode]{withMetaMode()}}), returns a
code expression (instead of evaluating that expression and returning the value).
}
\details{
If you wish to capture specific code inside of \code{expr} (e.g. ignore code
that has no meaning outside shiny, like \code{\link[=req]{req()}}), use \code{metaRender2()} in combination
with \code{metaExpr()}. When using \code{metaRender2()}, \code{expr} must return a \code{metaExpr()}.
}
\seealso{
\code{\link[=metaExpr]{metaExpr()}}
}
