% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.shinyMetaExpr}
\alias{print.shinyMetaExpr}
\title{Print method for shinymeta metaExpression}
\usage{
\method{print}{shinyMetaExpr}(x, formatter = format_tidy_code, ...)
}
\arguments{
\item{x}{an object of class shinyMetaExpr.}

\item{formatter}{a function that takes an unevaluated expression
and returns a character string.}

\item{...}{arguments passed along to [cat()].}
}
\description{
Print method for shinymeta metaExpression
}
